Added my own webscraper library
Sorry for leaving this repo empty! I tried to make a webscraper but I failed due to incompetence, after 3 more attempts to get it to work, this version passed both of my tests (a website that was auto-generated (like wix) (name of site withheld due to it being personally identifying) and cock.li which is much more rudimentary)

I'll admit, this code is awful and I am ashamed to put this code out on the internet in such an unrefined state but I think a commit is warranted after TWO months!!!

The code is "multi-threaded." I put this into quotation marks not because it isn't multi-threaded but because this is the first time I have ever used python threading and the system I made may be really bad!